# server/openapi.yaml
openapi: 3.0.3
info:
  title: AskFlow API
  description: |
    RESTful API for the AskFlow FAQ & Knowledge Base platform.

    ## Authentication
    Use an API key via the `X-API-Key` header to authenticate requests.
    Create API keys in the AskFlow admin panel under Settings > API Keys.

    ## AI & Chatbot Integration
    AskFlow is designed to integrate with chatbots and AI assistants:

    - **`POST /ask`** — Send a question, get the best matching answers (ideal for chatbots)
    - **`GET /context`** — Get all FAQ data for RAG/knowledge base population
    - **`GET /search`** — Full-text search with relevance ranking

    ### Supported Platforms
    - **Custom chatbots** — Use the /ask endpoint directly
    - **Botpress / Dialogflow / Rasa** — Use /ask as a webhook or custom action
    - **LangChain / LlamaIndex** — Use /context to populate vector stores
    - **Claude / ChatGPT tools** — Use /ask as a tool/function call

    ### Quick Start
    ```bash
    # 1. Get your API key from the admin panel
    # 2. Ask a question
    curl -X POST https://your-domain.com/api/v1/ask \
      -H "X-API-Key: faq_your-api-key" \
      -H "Content-Type: application/json" \
      -d '{"question": "How do I get started?"}'
    ```
  version: 1.0.0
  contact:
    name: AskFlow
servers:
  - url: /api/v1
    description: API v1
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    FAQItem:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
    FlatCategory:
      type: object
      properties:
        id:
          type: string
        product_id:
          type: string
        parent_id:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        sort_order:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /categories:
    get:
      summary: List all categories (flat)
      operationId: getCategories
      responses:
        '200':
          description: Flat list of all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlatCategory'
    post:
      summary: Create a category (use parent_id for nesting)
      description: |
        Create a category within a product. Use `parent_id` to nest categories
        at any depth (sub-sub-sub-subcategories are supported).
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, title]
              properties:
                product_id:
                  type: string
                  description: ID of the product this category belongs to
                parent_id:
                  type: string
                  nullable: true
                  description: ID of parent category (null for root categories)
                title:
                  type: string
                description:
                  type: string
                  default: ''
                icon:
                  type: string
                  default: ''
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatCategory'
  /categories/{id}:
    put:
      summary: Update a category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                icon:
                  type: string
                parent_id:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated category
    delete:
      summary: Delete a category (cascades to children)
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '404':
          description: Category not found
  /faq-items:
    get:
      summary: List FAQ items
      operationId: getFaqItems
      parameters:
        - name: category_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by category ID
      responses:
        '200':
          description: List of FAQ items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FAQItem'
    post:
      summary: Create a FAQ item
      operationId: createFaqItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, question, answer]
              properties:
                category_id:
                  type: string
                  description: ID of the category this FAQ belongs to
                question:
                  type: string
                answer:
                  type: string
                  description: Answer (supports Markdown)
      responses:
        '201':
          description: Created FAQ item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FAQItem'
  /faq-items/{id}:
    get:
      summary: Get a single FAQ item
      operationId: getFaqItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: FAQ item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FAQItem'
        '404':
          description: FAQ item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a FAQ item
      operationId: updateFaqItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                answer:
                  type: string
                category_id:
                  type: string
      responses:
        '200':
          description: Updated FAQ item
    delete:
      summary: Delete a FAQ item
      operationId: deleteFaqItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '404':
          description: FAQ item not found
...